Port the original `reproductions-2.cy.spec.js`
